+++ src/persona_chess/dataset/__init__.py
@@ -1,4 +1,4 @@
-from persona_chess.dataset.builder import build_move_examples
+from persona_chess.dataset.builder import build_move_examples, iter_move_examples
 from persona_chess.dataset.records import MOVE_EXAMPLE_SCHEMA, MoveExample
 from persona_chess.dataset.split import DatasetSplit, SplitConfig, split_examples
 
@@ -8,5 +8,6 @@ __all__ = [
     "MoveExample",
     "SplitConfig",
     "build_move_examples",
+    "iter_move_examples",
     "split_examples",
 ]